Multimodal_AI_Chart_Analyzer/requirements.txt
aiogram>=3.4.0
google-generativeai>=0.4.1
python-dotenv>=1.0.1
requests>=2.31.0
Pillow>=10.2.0
